refactor and group pool warnings into a single component
