+++ modules/pool/detail/components/warning/PoolDetailWarning.tsx
