+++ modules/pool/detail/components/warning/PoolStakeInFarmWarning.tsx
